requirements.txt
tencentcloud-sdk-python-intl-en>=3.0.0
cos-python-sdk-v5>=1.9.0
urllib3<2
